Add support for arbitrary data structures in MatchUnits
The purpose of this commit is to generalize the match units so
that the user can provide an implementation for the actual data
structures that are used. My particular use case is using the behavioral
model within a SystemC project. In order to obtain timing information we
need to provide our own data structures which interact with our memory
model. Another example use case would be changing to a different Trie
implementation which may perform better for particular patterns of
lookups.

Additionally, this change reduces the duplication in
match_units.{cpp,h}, factoring out the common logic into a templated
class, and the actual lookup logic into the separate
lookup_structure.{cpp,h} classes.