Add ability to retrieve a match entry from key (#213)
The Thrift RPC now exposes a bm_mt_get_entry_from_key method which can
be used to retrieve a match-action entry (along with the handle) by
providing the same match_key (and priority for ternary + range). The
handle can then be used to modify an entry, retrieve
counters... Although this means having to do 2 successive RPC calls, it
was easier to do than having to provide alternatives to all RPC methods
which currently require a handle.

The corresponding CLI command was added.