Add tentative support for parser errors and verify (#259)
P4_16 introduces error declarations. The core library defines a few
standard errors and each architecture can extend erros with additional
declarations. Parsers may 'raise' errors and architectures may make the
erros status available to controls.

In this pull request we try to add support for P4_16 declarations. It is
expected that the bmv2 compiler will assign a unique numerical value to
each error constant. This mapping (error name to value) will be made
available in the bmv2 JSON input. In the rest of the JSON file, the
errors are referenced only through their numerical value. However, for
parser verify statements, because the mapping is available, it is
possible to log a meaningful message using the error P4 name. The
mapping is also made available to architecture implementations, through
Context::get_error_codes and Switch::get_error_codes.

The bmv2 parser implementation supports (partially) the errors declared
in the P4 core library. Additional architecture-specific errors can be
raised through verify statements, which are now supported.

In the case where an architecture needs to make the error code available
to controls, it is the architecture implementation's responsibility to
copy the error code value to the appropriate metadata field.

Few changes were required to make this work. We use C++ exceptions to
manage P4 errors in the parser implementation, which proved convenient.

The JSON documentation was updated to include error support and the JSON
version was bumped up to 2.1.

This pull request also includes some minor cleanups in the parser code and in
the parser unit tests.