Add support for verify statements in bmv2 backend
With version 2.1, the bmv2 JSON supports declaring parser errors as well
as verify statements in the parser's body. See
https://github.com/p4lang/behavioral-model/pull/259/files#diff-418314966c57fea403ba29bf6b0793c7
for more information.

In this pull request, we add a visitor pass which assigns a numerical
value to each error declared in the P4 input. The mapping error name <->
value is added to the JSON under the 'errors' key. verify statements are
then handled when converting parser states to JSON, as per the bmv2 JSON
description.

Later on, we will be able to handle references to error constants in
expressions in controls. This will be done by substituting the numerical
value for the error constant in expressions, and otherwise using regular
bmv2 expressions. The error code produced by the parser will need to be
exposed as a metadata field though. So for now only references to error
constants in verify statements are supported.