Add support for enums in bmv2 backend (#247)
Starting with version 2.3, the bmv2 JSON supports declaring enums. See
https://github.com/p4lang/behavioral-model/pull/270 for more
information.

The bmv2 backend was already converting enums to integers (under certain
conditions). In this PR, we simply reteieve the mapping enums -> int and
dump it to the JSON file. In order to do this, the ConvertEnums midend
pass had to be slightly modified to give the backend access to the
produced mapping.