Support for stack_field type in expressions (#298)
The special stack_field type gives access to a field in the last valid
header instance of a stack. Until now, this was only supported in the
parser transition key. For the sake of uniformity it is now supported in
expressions and as an action parameter. This may go away in the future
if we adopt a more uniform scheme and treat "everything" as an
exception. In particular we should probably try to allow expressions in
parser transition keys.

The JSON documentation was updated to reflect this, and the version
number was bumped up to 2.6.