Add support for table initializers in JSON (#303)
This is done using the new 'entries' attribute in the JSON description
of a table. When this attribute is present the table becomes immutable:
JSON-provided entries cannot be deleted / modified and new entries
cannot be added.

The JSON documentation was updated to reflect this, and the version
number was bumped up to 2.7.