Add direct implementation of PI C API
This implementation runs in the same process as the bmv2 switch, as
opposed to the existing recommended implementation which is built on top
of the Thrift bmv2 service. This was added to avoid the Thrift
dependency when desired.

We provide 2 new flags for configure: --with-pi and
--with-thrift. --with-pi is used to enable compilation of the PI
implementation. Use --without-thrift to disable compilation of the code
that depends on Thrift. Of course if you disable Thrift, you will need
an alternative way of programming the dataplane. When Thrift is
disabled, unit tests are disabled and won't be run by make check.