Improve log messages for parser set operations (#313)
The log messages used to display the header id and field offset
(e.g. for destination field). It now displays the actual name of the
field.

This is definitely not the most efficient implementation, as we do a
name "lookup" every time, but it was the easiest to write. If there is a
need to increase performance even when logging is enabled, we can change
the implementaion to cache the name.