adds table entries support
Adds two IR nodes (IR::Entry and IR::EntriesList) as table properties

Adds the serialization of predefined entries to JSON (requires JSON
v2.7 or later) and a series of tests for table entries.

As part of the serialization, the match key type detection was
refactored to be common for generating table match keys and
determining the type of keysets in entries.

The typechecker needed convincing to accept action 'calls' in the
entry definition.
Action call parameters were being removed by the removeParameters
pass, which should not happen -- action data needs to be serialized
into the bmv2 json.

When disabling for IR::EntriesList, the type inference complains
... not sure why. So there are a couple of instances in the type
inference visitors where we check whether or not we are in an
EntriesList context and elide the type parameter checking. There
should be a cleaner way.

CONST ENTRIES grammar rule conflicts with CONST IDENTIFIER rule. For
now the only way this works is removing the CONST. We need to bring it
back.