adds documentation generation using doxygen
Documentation generation relies on the ax_prog_doxygen macros. By
default we generate HTML and PDF documentation which includes class
hierarchies generated using the 'dot' tool. The macros provide
enable/disable flags for different features of doxygen (see configure
--help). The documentation generation is disabled by default in the
bootstrap.sh script. Use --enable-doxygen-doc to enable. The README.md
has been updated with the required dependencies.

The configuration for doxygen is in the docs/doxygen directory. The
p4c_layout.xml file is used to customize the look of the HTML
documentation (this is still work in progress, as the default layout
is not very friendly).

Documentation is produced from source files (.def, .h, .cpp) and from
markdown (.md) files.  To ensure that introductory chapters are
produced in a pre-defined order, markdown files in the docs/doxygen
directory follow the convention xx_name.md, where xx is the ordinal
number of the chapter.

Still TODO:
* define chapters for documenting components of the compiler and hook
  them up in the navigation layout.
* more customization of the HTML documentation

Fixes #383