resolve shift/reduce conflict by moving annotations to the end of entry rule
Changes the grammar to rename actionInvocation to actionBiding and move
annotations for entries at the end of the statement, such that we avoid
a shift/reduce conflict in the grammar.

Also tried splitting assignmentOrMethodCallStatement into
assignmentStatement and methodCallStatement that is just transforming
an actionBinding (MethodCallExpression) into a Statement, but that creates
more issues in the grammar.

Fiddled with the tests to move annotations at the end ... so happened that
I had to pick the test with annotations to test all error conditions :(