Add support for writing P4Runtime in text format
This is an alternative to the JSON format for debugging / testing /
manual editing. Because we now support 3 formats: binary, json and text,
it made sense to update the command line options support to use a
choice-type argument instead of having a flag for each format. Therefore
this introduces '--p4runtime-format' and deprecates
'--p4runtime-as-json'.