Add better support for VL / varbit fields
I have tried to add support for variable-length fields (aka varbit
fields in P4_16) which is consistent with the P4_16 spec.
The following has been added:
 * 3 core action primitives available for all bmv2 target architectures:
   assign, assign_VL and assign_header. These essentially deprecate
   modify_field and copy_header for simple_switch.
 * the extract_VL parser operation which takes 2 parameters, the
   destination header and a length expression, and behaves like the
   2-parameter version of the P4_16 extract built-in.

These changes are backward-compatible and bmv2 still supports
P4_14-style JSON, where the length expression is a property of the
variable-length header and not a parameter to extract.

Very few runtime checks are performed for varbit assignment, so we rely
on the compiler to perform all necessary checks at compile-time.

The JSON documentation was updated to reflect this, and the version
number was bumped up to 2.9