replace the use of typeMap->constants for tracking compile time constants
This commit eliminates the use of the typeMap->constants to keep track
of compile time constants, and relies on the IR nodes to know whether
they are constants themselves. There are a couple of instances
(Members and PathExpressions) where the compile time constantness is
defined by the type declaration. In this case we add a method just to
those nodes to give them the ability to be set as constants.