(Part 1) - Add an optional dependency on libboost-iostreams
The main refactoring taking place in this PR is to port the P4-14 and
P4-16 lexers and parsers to use the C++ modes of Flex and Bison. In C++
mode, Flex generates a lexer that reads from a `std::istream` object
instead of a file descriptor. That produces a bit of an impedance
mismatch, because typically before we parse anything we run it through
the C preprocessor via `popen()`, and `popen()` uses C `FILE*`-based I/O
rather than iostreams.

The C++-11 standard library does not include a `FILE*` wrapper that
produces an `std::istream`. If we don't want to write and maintain one
ourselves (and I'd argue we don't) that leaves us with three options:

1. Stop using `popen()` and implement the subset of C preprocessor
   features we need directly.

2. Read the entire file in and wrap it in a `std::stringstream` or
   similar. This obviously isn't as efficient as it could be.

3. Use Boost's IOStreams library to do the job more efficiently.

In the large term, my view is that #1 is the best choice, but in the
short run #3 seems like a win with no real downsides other than adding
another dependency. (It's only an additional dependency on Ubuntu, which
splits Boost into a number of small packages. On macOS there's actually
no change.)

In the interest of getting this PR landed ASAP, I've actually
implemented both #2 and #3, so the Boost IOStreams dependency is
optional. I would vastly prefer to make it mandatory, though.