(Part 2) - Prepare to remove unused CompilerOptions parameter to IR::V1Program()
IR::V1Program's constructor took a CompilerOptions parameter that it
doesn't actually do anything with, and its implementation was the sole
contents of `frontends/p4-14/intrinsic.cpp`. This patch cleans things up
a little bit by moving the implementation to `ir/v1.cpp` (which seems to
be its logical home since the declaration is in `ir/v1.def`), removing
the now-empty `intrinsic.cpp`, and eliminating the CompilerOptions
parameter. I retained it as an overload so that this commit can build
separately, but that overload is removed later in this PR.