(Part 4) - Add a parser_error() overload that matches Bison's C++ API (and add comments)
The existing implementation of `parser_error()` does a lot of work to
infer the location of the message its printing; it assumes that the
parser itself can't provide the necessary location information. For the
C++ Bison parser, that's not true, so this patch adds a `parser_error()`
overload that just accepts the location information from the parser.

This implementation mostly produces the same results. I did see one case
in our tests (which is among the updated testcases) where the result is
a bit better.

We will eventually be able to remove the old implementation, but the IR
generator uses this code also, and I've avoided making any changes to it
in this PR.