support for a binary distribution (#647)
* allow loading of configuration files when no source is specified

In order to print the available targets, we need to allow that the
configuration files can be loaded even when a source file is not
specified. In addition, all the arguments checking should happen only
after we figured out whether we've been asked for help.

* generalize p4c to find its own paths

Remove the autoconf variables and base p4c on its current directory.
Handle both the developer installation and the binary installation.

Fixes #501

Remove the current directory from the path, and provide an explicit
variable to find the compilers. This avoids issues with running
obsolete versions of the compiler.

Fixes #502

* generate p4c using autoconf to handle the case when datadir is specified explicitly

* catch exceptions when not all artifacts are installed

And remove empty lines when no output is generated by subprocesses.