adds build type env variable to driver script
The driver is intended to invoke all the required tools for each backend.
It is used in two scenarios, developer and installed version, and the
tool binaries may be located in different directories. We set an env
variable to identify in which scenario is the script invoked, and this
env var can be used in the configuration file to set the correct
paths.