Better STF check to test if simple_switch is ready
The previous code had a simple 2 second timeout which was an issue in
some cases (very slow system, "warm-up" time for libtool wrapper script
if a non-installed simple_switch is used, ...). The new timeout is much
larger but we also periodically check whether the bmv2 process is still
running to ensure we are not waiting for nothing.