adjusts the paths for the include files to the env variable set by p4c driver
Install and developer builds provide different paths to the include files.

In the current configuration, we hardcode the p4include path to some
directory decided at configuration time, which makes the compiler
non-relocatable. However, this issue has been addressed in the driver,
which sets environment variables for both the P4 14 and 16 include
paths. With this change, the compiler no longer needs to hardcode the
path.