In P4Runtime, reject anonymous table properties
Rather than generating a default name. Until now the default name was
the name of the property itself (e.g. 'implementation'). This is a
problem when multiple tables have anonymous properties, but also because
1) the name can be surprising to the control-plane programmer, and 2)
backends can generate their own names for these which may not match
(which is what the bmv2 backend is doing). Taking that into account, it
seems better to just raise an error in the P4RuntimeSerializer.