Support 64-bit device id
Some users have expressed interest in being able to use 64-bit device
ids with bmv2. The biggest issue with this change is that the format for
notification messages (ageing, learning, port) has changes to accomodate
the larger device ids. As a consequence external tools subscribing to
these messages directly need to be updated.