Fix use of unexported GC library variables (#849)
* Fix use of unexported GC library variables (issue #844 )

The variable GC_start_call_back was being used directly, but is not part of the
official library interface, ie marked with GC_API. This means that some systems
may not have this call available (notably Windows/MSYS2). GC version 7.2.0
introduced an API GC_set_start_callback() for setting this callback. This
commit uses that call and places a minimum version of 7.2.0 for p4c builds.

Any manipulation of GC_print_stats has been bypassed on Windows systems,
which is also an unofficial and unavailable API on that system. There appears
to be no official way of supporting that.

Finally, the extern declaration for GC_get_heap_size was removed as it is
part of the official APIs.

Signed-off-by: David George <david.george@netronome.com>
Acked-by: Rando Wiesemann <rando.wiesemann@netronome.com>

* Rework GC_print_stats availability check

Previously the GC_print_stats detection was based on platform C defines. It
now uses CMake check_function_exists() and a CMake define HAVE_GC_PRINT_STATS.
This is both neater and more robust.

Signed-off-by: David George <david.george@netronome.com>
Acked-by: Rando Wiesemann <rando.wiesemann@netronome.com>

* Reinstate removed comment

* Improve library handling for check_function_exists

Use LIBGC_LIBRARIES and not gc to cater for non-standard installations. Also
make sure changes to CMAKE_REQUIRED_LIBRARIES are only in place for the
check_function_exists calls.