Add ability to serialize static table entries (#912)
* Add ability to serialize static table entries

In the case where a P4_16 program provides compile-time table entries,
it is convenient for targets in general to have access to P4Runtime
messages which encode these static entries. This adds support for
serializing all static entries in a program as a P4Runtime.WriteRequest
message. This message can then be sent to a P4 target along with the
rest of the compiler output, and the message can be processed just like
runtime control-plane messages.

When using the bmv2 backend, the '--p4runtime-entries-file' command-line
option can be used to generate the message and write it to the provided
file. The format is determined by '--p4runtime-format'.

This commit was a good opportunity to update the submodule ref pointer
to p4lang/PI, which provides p4info.proto and p4runtime.proto.

* Print warning if --p4runtime-entries-file is used alone

Without --p4runtime-file...

* Add unit test for P4Runtime static entries converter