Fix some of the failing STF tests described in #961
These failures all are due to bmv2 sending packets out of port 0 for a
P4_14 program that never assigns a value to the egress_spec intrinsic
metadata field.

Chris Dodd suggested changing the programs so that if they do not
explicitly assign a value to egress_spec, then they drop the packet.
This makes the existing STF tests for these programs pass, so these
tests appear to have been written with this behavior in mind.