streamlines option processing
Since different backends may want to add options that are proprietary
we refactor the option processing such that they can add these options
as part of different argument groups.

Also, segregated the processing of generic options from the options that
require the source file to be present. The source dependent options are
typically backend dependent, so it is expected that backends will set them
in the subclass, just before the commands are run, rather than part of the
"configuration".