Add new core assignment primitives for stacks
Added assign_header_stack for header stack assignment and
assign_union_stack for header union stack assignment. Stack copies are
valid in P4_16. It is not possible for the compiler to generate
equivalent code by copying each stack element individually as stack have
internal state (the next "pointer"). Having dedicated bmv2 primitives is
also more efficient.

The JSON documentation was updated to reflect this change, and the
version number was bumped up to 2.17