[Experimental] value_set as first class type
This PR introduced value_set as a first class type for the parser
sub-language. The value_set type behaves like a set, but the compiler
should make sure a variable of type value_set can only be used in the
parser sub-language in the select cases. It should be illegal to use
value_set anywhere in the control block.