Fix locking model in BMI
The previous code was broken. In particular, removing a port while
sending to / receiving from that port could have lead to a race
condition. Copying the FD set in the run_select function also wasn't
protected with a lock, but port_add & port_remove modify the set.

We use a RW mutex to avoid taking a performance hit and ensure that send
& receive do not block each other.