Add egress timestamp to v1model's metadata fields
This is to enable the implementation of INT with v1model. This metadata
field does exist in PSA, but until PSA becomes mainstream, it is nice to
have access to the same metadata in v1model.

The egress timestamp is the "time at which the packet began processing
in the egress parser".