Make P4_16 stack implementation the default
As previously announced, we are making the P4_16-conformant
implementation for header stacks the default in prevision of the bmv2
1.11.0 release.

As a remainder, the new implementation differs from the old one in the
following ways: push_front and pop_front shift the entire stack (not
just until the next index), and pushed headers are marked as invalid
instead of valid.

The legacy implementation can be toggled by passing the
`--disable-WP4-16-stacks` flag to configure.

We recommend that exsiting P4_14 compilers inject an instruction to make
pushed header valid if they want to achieve a similar behavior as the
legacy implementation.

Fixes #557