adds support for saturating arithmetic in P4_16
We add support for |+| and |-| in the frontend of the compiler, as
well as bmv2 backend support. This includes defining two binary
operation IR nodes for saturating addition and subtraction, and
generating the json for bmv2 as defined in
https://github.com/p4lang/behavioral-model/pull/590. For that, we
update the json version to 2.18.

We also added a test that ensures that saturating arithmetic properly
bounds the min and max values for signed and unsigned integer
operations.

Still to do:
- translate P4_14 saturating operations to the new IR expression nodes.

Fixes #1229