Use correct ID prefixes in P4Runtime serializer
The different values for the ID prefix (first byte of every ID) are now
defined in p4info.proto. In this commit, we switch to using these values
when generating the P4Info message. One notable difference is that the
direct & indirect versions of counters / meters now use a different
prefix value for their ids.