Move p4RuntimeSerializer invocation to right after frontend
For bmv2 and p4test backends. It is important to run the serializer
before some of the midend passes that could have an impact on the
contents of the P4Info message - such as the EliminateTuples pass.

Currently, the serializer seems to have a dependency on a few passes
usually run in the midend, which is why it has its own pass manager.

Note that if a given backend later runs some transformation pass like
EliminateTuples, it is the responsibility of this backend to provide the
P4Runtime server implementation with the necessary information to do
data transformation.