A framework for P4Info generation for arbitrary architectures
More architectures are being developped and there is a desire to be able
to use P4Runtime with them. P4Runtime supports standard architectures
(v1model & PSA) natively and support architecture-specific extensions
(table properties & externs). In order to enable using P4Runtime with
vendor-specific architectures, we define a framework that can be used to
implement P4Info generation for these architectures more easily. v1model
& PSA also use this framework, but support for these standard
architectures is supported natively by p4c, while vendor-specific
architectures must be supported in the vendor's backend.

The goal of the framework is to reduce code redundancy. A
P4RuntimeArchHandlerIface implementation can be registered for each
architecture. The P4RuntimeSerializer singleton object takes care of all
the common logic, and calls the overridden P4RuntimeArchHandlerIface
when needed for architecture-specific processing.

There are still some minor v1model references in P4RuntimeSerializer
that will be solved in a future commit. These are minor and are
dependent on some fixes in p4info.proto.

P4Info generation for PSA is not supported yet (it will be in a future
commit) but most of the v1model code should be re-usable.