bugfix: default_action translation to p4-16
During the P4-14 to 16 translation, when a table has an attached direct_counter
or direct meter, the translation inserts a counter.count() or meter.execute()
method call to each of the action in the table.

As a result, if an action is used in the default_action property of the table,
the default_action has changed from the original unmodified action to the
updated action with counter.count() or meter.execute().

The current P4-14 to 16 translation has a bug which does not update the
default_action property in the table, and the wrong default_action is used when
a direct_counter or meter present.