Add support for PSA in p4RuntimeSerializer
Because PSA and v1model share many similarities (not to mention that
they are both considered standard architectures in the context of
P4Runtime), we tried to unify the code for the two architectures as much
as possible using meta-programming. We define a "traits" class for each
extern type to handle the difference in naming (table property names,
extern type names, extern method parameter names). One major difference
between the two architectures is digest: v1model uses an extern method
while PSA defines an extern type.

We added a new gtest unit test for the PSA code, very similar to the
ones we already have for v1model. Unfortunately PSA requires a lot of
boiler-plate code so the test is pretty verbose. We may want to find a
better solution in the future. The unit test covers digest, since as
pointed out before this is the major difference with v1model.