Support standard_metadata.parser_error field in SimpleSwitch
v1model now includes a new standard metadata field that is used to
communicate the parser error code to the control blocks. This commit
makes sure that the code is written to the field after the parser
executes. The change itself is very small but we include a unit test for
SimpleSwitch (with P4 & JSON files).

Fixes #446