Add 2 hash functors to the Bmv2 core library
"round_robin" and "round_robin_consistent"
The first one keeps an internal counter which is incremented by 1 and
returned every time the functor is called. The consistent version keeps
an internal map and guarantees consistent results when the functor is
called multiple times with the same input.
This can be useful when the hash is used for an action selector, as a
way to iterate through all possible members and obtain all possible
outputs.

Fixes #641