add support for uint64_t constants in IR::Constant (for 32-bit)
Casting to long on 32-bit machines does not achieve the expected
result that the value is 64-bit. This PR adds support for uint64_t in
IR::Constant, by expanding the set of constructors with the
almost all used integer data types (exception is unsigned long because the
ir parser generator does not handle types that have two qualifiers).

TODO: deprecate the use of asLong throughout the code base. Right now,
it is used only in v1parser.ypp, and as far as I can tell, only for
sizes that should be well within 32-bits (ids, bit-widths, etc.). I do
recommend a review of the code, and replacement with asUnsigned().

Added a gtest to check different values for all the types -- which is really
what forced me to the solution adding all constructors!