warnings
Using generic names like COPY outside name spaces is bound to sooner or
later crash with an external library. Since COPY and IGNORE are macros
defined only for the ir-tree-macros benefit, I renamed them locally.

We seem to have a template of passing the typeMap to different passes
even though it is not used in the pass. I have removed a number of these
as well as a number of other inconsistencies.

Another big source of warnings seem to be unused parameters in the
google::protobuf objects. I disabled these warnings using pragma diagnostics
as we're not going to go into that code and fix them.

These were reported on Mac OS with Apple LLVM version 9.1.0 (clang-902.0.39.2).