Remove PI submodule and use p4runtime instead
The P4Runtime specification (including the Protobuf definition files) is
no longer hosted on p4lang/PI but on p4lang/p4runtime, so we updated the
submodule to refelect that.
We need a third-party Protobuf file to compile P4Runtime
(google/rpc/status.proto). Rather than pull-in the entire googleapis
repository (100+MB) where this file is hosted, we add a copy of the file
here.