add support for versioning p4c backends
We add support for versioning different p4c backends independently of
each other. Currently we specify a global version in the root
CMakeLists.txt file, which will be reported by the p4c driver.
Individual backends get a configurable version string that is set at
configuration time based on macros derived from the global
version. However, each backend can customize the format of the
version, and see the example in the psa backend that hardcodes the
version to 0.0.1 and uses only the git SHA.

Alternative solution to #1491.