Added support for free-form annotations (untested)
This adds support for free-form annotations, wherein annotation bodies
may be any sequence of tokens containing balanced parentheses. At the
same time, this introduces annotation checking, to ensure predefined
annotations are used correctly.

On the initial parsing pass, annotation bodies are parsed as a sequence
of tokens in a new field, `Annotation::body`. After the initial parse,
`Annotation::expr` and `Annotation::kv` are empty.

Early in the front end, a rewriting pass parses the bodies of
spec-defined annotations, converting them into an expression list (which
populates `Annotation::expr`) or a key-value list (`Annotation::kv`).
This has the new side-effect of rejecting annotations with mismatched
bodies. For instance, having an expression list where a key-value list
is expected, or having a no-argument annotation with a non-empty body.

As of v1.0.0-draft of the spec, annotations are as follows:
  * `@optional` with an empty body (§18.2.1)
  * `@tableonly` with an empty body (§18.2.2)
  * `@defaultonly` with an empty body (§18.2.2)
  * `@name` with a string literal (§18.2.3)
  * `@hidden` with an empty body (§18.2.3)
  * `@atomic` with an empty body (§18.2.4)
  * `@pkginfo` with a key-value list (§18.3)

There is also a `@length` annotation that `p4c` supports, but is not
mentioned in the spec. This annotation takes an integer literal.

Unrecognized annotations are left untouched. For any target-specific
annotations, back ends may do a similar rewriting pass as needed.