Add stub for _pi_act_prof_grp_set_mbrs
This is a new PI API to set action profile group membership with a
single call. We currently do not implement this method for bmv2, so
_pi_act_prof_api_support always returns
PI_ACT_PROF_API_SUPPORT_GRP_ADD_AND_REMOVE_MBR. This is fine because the
P4Runtime server implementation on p4lang/PI supports both programming
methods.