Populate Documentation fields and PkgInfo in P4Info
The documentation information comes from the '@brief' and '@description'
annotations. The PkgInfo message is populated based on the annotations
on the P4 package declaration.

To reduce code duplication, a new setPreamble helper function is added,
which can be used to set all the fields (including documentation and
generic annotations) in the preamble for all P4Info entities.

Fixes #1170