error reporting (#1646)
* fix warnings and name overload

Fixes a couple of warnings, and in particular the one in
parserControlFlow.cpp looks like a genuine error, where the variable
is redefined in each iteration.

* add support for typed errors

Addresses the issue of repeated error and warning messages when
compiler passes are repeated, by introducing error/warning types, and
an error catalog. Error types classify the errors and impose a format
that allows the compiler to automatically filter repeated
messages. The filtering is based on the type of error and the source
code location of the object that reports the error. Thus, it allows
multiple error types per source code line, and ensures that only one
error is reported even if the message is raised multiple times. Error
codes and formats are defined in `lib/error_catalog.[h,cpp]`. Backends
can extend the codes and formats as needed (and they are encouraged to
do so).

To ease the transition to typed errors and warnings, free form
messages whose first argument (`%1%`) is an `IR::Node` (or more
precisely a class that implements `Util::IHasSourceInfo` interface),
will be converted automatically to typed errors that use the format
argument of the message rather than the error catalog predefined
formats.

Also introduces the `FATAL_ERROR` macro to print a message and raise
an CompilationError exception.

And finally adds documentation on how to use to the coding standards.

* convert bmv2 errors to typed errors

* convert to typed errors and warnings in the frontend

* revert robotic change for arguments that are strings

* refactor the error catalog

Refactor the error catalog into a singleton class and keep user format
messages combined with a error/warning title.

Define additional classes of warnings to cover the majority of warnings
in the frontend and the three open-source backends.

* introduce error and warning types.

* update reference outputs

* addresses reviewer comments

More cleanup on the error messages:
- remove duplicated printing of the leading object
- add ERR_TYPE_ERROR as an error type to enable tracking all uses of fields
- add ERR_UNSUPPORTED_ON_TARGET error type
- fix spacing
- rebased on latest master

* allow errors for all occurrences of a non-defined field