Handle complex headers (with nested structs) in P4Info generation (#1708)
The P4Info generation code now handles complex headers by flattening
them. This is independent from the FlattenHeaders pass, which is a
midend pass. This code is much simpler and has no dependency on non
frontend pass. It simply performs a "local" flattening of a given
Type_Header, without updating any program references. It also preserves
the fully-qualified name using a @name annotation (unlike the
FlattenHeaders pass), along with annotations of the nested
fields. Because of the local nature of this flattening, it seemed silly
to write an actual pass (I tried and it was clumsy, with a call to
visitAgain).

The flattening is performed wherever the p4RuntimeSerializer has to deal
with header types, so only 2 places: for controller headers and in the
typeSpecConverter.