Support --load-modules option for simple_switch_grpc
Since this is useful for many targets (simple_switch, simple_switch_grpc
and possibly psa_switch) and has been around for a while, I decided to
move support for --load-modules to the core bmv2 library, by adding a
new reference target-specific option parser
(`TargetParserBasicWithDynModules`) that targets can use directly. The
reason why this functionality did not find its way into the main bmv2
option parser is because it requires the target to be linked with
-rdynamic, so I believe it is better for the target to choose whether
this functionality is required or not. Not moving it to the main option
parser also guarantees backward-compatibility for simple_switch users.

Fixes #719