Add option to STF "expect" command for stricter checking of output packets
Before and after this change, the default behavior for the STF
"expect" command is to check that the actual output packet from BMv2
is at least as long as the one specified in the "expect" statement,
and all bytes in the "expect" packet data are present at the beginning
of the actual output packet.  If the actual output packet has
additional data after that, the test still passes.

With this change, if you add a "$" character at the end of the packet
data in an "expect" statement, the test will fail in all cases where
"expect" fails today, plus it will also fail if the actual output
packet is longer than the one specified in the "expect" statement.
This enables slightly more strict checking of actual output packets.