Detect duplicate names in P4Info
See issue #1803

We iterate over the P4Info after generating it to look for duplicate
names, and generate an error if we found one. Unfortunately we cannot
rely on the SymbolTable for duplicate detection, because exsting code
relies heavily on the ability to call add() multiple times with the same
name. For example top-level actions are inlined by the compiler in all
the controls using them and we end up calling add() for each copy. In
this case, it is not considered an error because these "duplicate" calls
actually refer to the same object as far as the control-plane is
concerned and we do not end up with duplicates in the P4Info message.

A negative test case is added to p4_16_errors. This required slightly
changing the testing framework so that programs under p4_16_errors are
compiled with --p4runtime-files.