Make changes to 3 PSA tests that are causing the build to fail
Two of them are passing now, and should be removed from the XFAIL list.

One of them was written in a way that it was passing, but it should
not have been.  Updated the STF test so it fails.  A separate PR on
behavioral-model code is required to make that test pass.