Remove unnecessary PSA test program
Early in PSA specification development, there was the proposal that a
single table could have multiple direct counters associated with it,
as demonstrated by this test program.  That generality was later
removed before PSA v1.0 was published, since such generality seemed
unlikely to be implemented, and unlikely to be needed by anyone.

All of the other functionality of psa-counter5.p4 is covered by
the test program psa-counter4.p4