Update several PSA test programs to avoid the effects of issue 1918
Using the slice operator in a PSA program action is one language
feature that causes issue 1918 to happen.  By avoiding that feature in
an action, and any other language feature that causes a mismatch
between the actions added and the actions later looked up in the PSA
back end code, the BMv2 JSON files are correct (or at least not wrong
for the reasons specific to issue 1918).