Fix Travis CI by removing python-pip from Dockerfile deps
Instead install pip with get-pip.py
(https://pip.pypa.io/en/stable/installing/).

python-pip seems to be broken on Ubuntu 16.04. The installation of pip
itself is failing (it seems that it fails to reach PyPI, which may be
specific to the Travis environment, which causes a backtrace).

We also remove Tenjin from the python dependencies, since it is not used
by p4lang/p4c.