Structs, headers and tuples initialisation support (#762) (#2017)
Added support for struct/header/tuple initialisation with empty tuple.

Types that are supported as leaf fields are int, bool and bit.
Int and bit type fields will be initialized to zero, bool to false.
If any other type is used as a lief field compiler will report error.

Frontnend reorganized, Struct initializer now processes declaration initializers
along with assignment statemants, which is the cause of changes in expected oputputs
of some tests.