add a compiler option to disable specific annotations/pragmas
Add the --disable-annotations option to disable annotations by name.
It accepts a comma separated list of annotations that will not be
added in the IR tree. A warning will be printed that the annotation
was explicitly disabled. We do print the warning for every instance of
the annotation, which may be an overkill, however, it is intentional,
so that users can make sure that a particular annotation was not taken
into account.

The option only disable annotations present in the source
file. Compiler passes may continue to add annotations and process them
as before. Note that users may be able to disable annotations that may
cripple the code generation -- backend or control-plane. That's
another reason why it is important to print those warnings!