Move to Python3 (#2087)
Converts scripts to Python3

Moves all python scripts in the p4c repository to Python3, except for cpplint.

We should consider moving to https://github.com/cpplint/cpplint to avoid maintaining cpplint. However, this requires adapting all the linting rules to the new version of cpplint, which is orthogonal to this PR.

We still install Python2 dependencies so that we don't break downstream projects. While I do not have a crystal ball to know when they are transitioned, we should put a deadline for these projects to transition and remove the Python2 dependencies.