Support lvalue expressions
The bmv2 expression engine can now returns references to bmv2 primitive
objects (Data, Header, ...) when the expression being evaluated is an
lvalue. This means that we can now handle assignments to complex lvalue
expressions which do not statically resolve to a field / header /
union. One example would be an assignment to a header stack field
accessed through a runtime index.

Unit tests were added, although at the moment they do not support all
the possible cases (e.g. expressions which evaluate to a header union).

The JSON version number was bumped up to 2.23. Even though no additions
were made to the format, we now support additional cases and the
documentation now includes an example of a "complex" assignment
statement.

Fixes #838