Correctly parse action references in const table entries
This pull request fixes a discrepancy between the language
specification and `p4c` related to the syntax for action references in
const table entries.

The grammar in the langauge specification allows action references to
be given without parentheses:

```
entry
    : keysetExpression ':' actionRef optAnnotations ';'
    ;

actionRef
    : optAnnotations prefixedNonTypeName
    | optAnnotations prefixedNonTypeName '(' argumentList ')'
    ;
```

However, the grammar in the `p4c` parser required all action
references to be given with parentheses:

```
entry
    : keysetExpression ":" actionBinding optAnnotations ";"
        { ... }
    ;

actionBinding
    : lvalue "(" argumentList ")"
        { ... }
    | lvalue "<" typeArgumentList ">" "(" argumentList ")"
        { ... }
    ;
```

This pull request adds a new clause to the `actionBinding` rule to
handle actions without parentheses, as well as a unit test
(`table-entries-no-arg-actions.p4`) that checks whether such action
references are correctly parsed.