Structs, headers and tuples initialization support (#762) (#2017)
Struct, header or tuple can be initialized automatically with a default value of
the suitable type using the syntax `...` or using a mix of explicit
values and default values by using the notation `...` in a list
expression initializer(and struct initializer expression for structs).

Mix of explicit values and default values can also be used for assignment.

For `struct', `tuple` and `header` types the default values are `struct`,
`tuple` and `header` respectively, where each field has the default value
of the suitable field type -- if all such default values are defined.

Default value for `header` type could be easily changed to invalid, as proposed
in p4-spec pull request #797, when a new literal for initializing `header` to
invalid is introduced.

Header stacks and `header_union` are not supported as a leaf fields of structs
and tuples if using this kind of initialization, because in current version of
spec there are no constructs to initialize these types.
If they are used as a leaf fields compiler will report an error.

Frontnend reorganized, Struct initializer now processes declaration initializers
along with assignment statemants, which is the cause of changes in expected outputs
of some tests.