Fix type_info generation
P4Runtime v1 resticts the type of some fields (match fields, action
parameters, controller metadata fields) to unsigned bitstrings. When a
user-defined type is used for one of these, P4Info will use the
underlying type, but will also populate a type_name field in the
Protobuf message so that P4Info consummers can be aware of the
user-defined type. The type_name can then be looked-up in the type_info
"database" of types. This commit fixes an issue which caused type_info
not to be populated correctly. It also makes sure that type_name is set
properly for counter / meter / register indices.

Fixes #2239