Ensure that ID prefixes are correct in P4Info when '@id' is used (#2243)
* Ensure that ID prefixes are correct in P4Info when '@id' is used

As per the P4Runtime specification, when the '@id' annotation is used:
  * if a prefix is missing (i.e. it is 0), assign the correct prefix
    based on the resource type.
  * if a prefix is present (i.e. not 0), ensure that the provided value
    is correct for the resource type - this is not part of P4Runtime
    v1.1, but is being added to the unreleased v1.2 specification.

Fixes #2241

* Use 24 bits worth of hash result when generating P4Info ids

For legacy reasons which no longer apply, we were only using 16
bits. That means that we can now have more than 65K P4 objects of a
given type (if someone ever needs that many...). Unfortunately a lot of
test golden outputs have changed as well.