Fix bmv2 default parser transitions
The bmv2 backend was still using a very old format (from 2016), and
while it is still supported by bmv2 for backwards-compatibility reasons,
it is confusing to people who compare the generated JSON files to the
bmv2 JSON format spec.

We also update JSON_MINOR_VERSION to the latest version.

Fixes #2311