Clarified an STF test case with a too-short packet
The pseudo-code for `packet_in.extract` P4 Language Specification
implies that the packet buffer should be unmodified if the packet is
too short. The `parser_error-bmv2.stf` test checks this case. But the
`expect` statement was potentially confusing, since STF checks that
the prefix of the emitted packet matches the bitstring.

This commit clarifies the test case, by explicitly listing the
original packet (which is not extracted) and marking the end of the
packet with `$`.